stage0: use yaml to parse pod manifest;
This enables us to write pod manifests in yaml, which is more easy to read and write for a human.
Since json is a subset of yaml, this change is completely backward compatible.
Yaml parsing is implemented using the ghodss/yaml wrapper, so that there is no need to update appc/spec with yaml-unmarshaler interfaces. (types.SemVer and types.Hash have special json handling)